% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg2mm.R
\name{hg2mm}
\alias{hg2mm}
\title{Convert between human and mouse orthologs using a variety of methods}
\usage{
hg2mm(
  genes,
  query = c("human", "mouse"),
  target = c("mouse", "human"),
  method = c("one2one", "unique_query", "unique_target", "all")
)
}
\arguments{
\item{genes}{the genes you wish to convert}

\item{query}{the species of your query genes}

\item{target}{the species whose orthologs you wish to convert to}

\item{method}{the method you wish to use to return the appropriate orthologs}
}
\value{
a tibble giving each of your genes and the corresponding orthologs in the other species
}
\description{
Convert between human and mouse orthologs using a variety of methods
}
\examples{
hg2mm(c("WDR53", "RETNLB"), query = 'human', target = 'mouse', method = 'one2one')
}
